+++ src/packet/encap.rs
@@ -9,6 +9,7 @@
 //! [encap]: <https://github.com/riscv-non-isa/e-trace-encap/>
 
 use super::decoder::{self, Decode, Decoder};
+use super::encoder::{Encode, Encoder};
 use super::{payload, unit, Error};
 
 /// RISC-V Packet Encapsulation
@@ -16,7 +17,7 @@ use super::{payload, unit, Error};
 /// This datatype represents a "Packet Encapsulation" as describes in Chapter 2
 /// of the Encapsulation specification.
 #[derive(Clone, Debug, PartialEq)]
-pub enum Packet<P> {
+pub enum Packet<P = payload::Payload> {
     NullIdle { flow: u8 },
     NullAlign { flow: u8 },
     Normal(Normal<P>),
@@ -57,6 +58,35 @@ impl<P> From<Normal<P>> for Packet<P> {
     }
 }
 
+impl<'d, U> TryFrom<Packet<decoder::Scoped<'_, 'd, U>>>
+    for Packet<payload::Payload<U::IOptions, U::DOptions>>
+where
+    U: unit::Unit,
+    U::IOptions: Encode<'d, U>,
+    U::DOptions: Encode<'d, U>,
+{
+    type Error = Error;
+
+    fn try_from(packet: Packet<decoder::Scoped<'_, 'd, U>>) -> Result<Self, Self::Error> {
+        match packet {
+            Packet::NullIdle { flow } => Ok(Self::NullIdle { flow }),
+            Packet::NullAlign { flow } => Ok(Self::NullAlign { flow }),
+            Packet::Normal(p) => p.try_into().map(Self::Normal),
+        }
+    }
+}
+
+impl<'d, U> Decode<'_, 'd, U> for Packet<payload::Payload<U::IOptions, U::DOptions>>
+where
+    U: unit::Unit,
+    U::IOptions: Encode<'d, U>,
+    U::DOptions: Encode<'d, U>,
+{
+    fn decode(decoder: &mut Decoder<'d, U>) -> Result<Self, Error> {
+        Packet::<decoder::Scoped<_>>::decode(decoder).and_then(TryFrom::try_from)
+    }
+}
+
 impl<'a, 'd, U> Decode<'a, 'd, U> for Packet<decoder::Scoped<'a, 'd, U>> {
     fn decode(decoder: &'a mut Decoder<'d, U>) -> Result<Self, Error> {
         if decoder.bytes_left() == 0 {
@@ -71,23 +101,46 @@ impl<'a, 'd, U> Decode<'a, 'd, U> for Packet<decoder::Scoped<'a, 'd, U>> {
 
         match core::num::NonZeroU8::new(length) {
             Some(length) => {
-                let src_id = decoder.read_bits(decoder.hart_index_width())?;
+                let src_id_width = decoder.hart_index_width();
+                let timestamp_width = decoder.timestamp_width();
+                let length = usize::from(length.get())
+                    + usize::from(src_id_width >> 3)
+                    + usize::from(timestamp_width);
+
+                let mut payload = decoder::Scoped::new(decoder, length)?;
+                let src_id = payload.decoder_mut().read_bits(src_id_width)?;
                 let timestamp = extend
-                    .then(|| decoder.read_bits(8 * decoder.timestamp_width()))
+                    .then(|| payload.decoder_mut().read_bits(8 * timestamp_width))
                     .transpose()?;
-                decoder::Scoped::new(decoder, length.get().into()).map(|payload| {
-                    Normal {
-                        flow,
-                        src_id,
-                        timestamp,
-                        payload,
-                    }
-                    .into()
-                })
+                Ok(Normal {
+                    flow,
+                    src_id,
+                    timestamp,
+                    payload,
+                }
+                .into())
             }
             _ if extend => Ok(Self::NullAlign { flow }),
             _ => Ok(Self::NullIdle { flow }),
         }
+    }
+}
+
+impl<'d, U, P> Encode<'d, U> for Packet<P>
+where
+    U: unit::Unit,
+    Normal<P>: Encode<'d, U>,
+{
+    fn encode(&self, encoder: &mut Encoder<'d, U>) -> Result<(), Error> {
+        let (flow, extend) = match self {
+            Self::NullIdle { flow } => (flow, 0x00),
+            Self::NullAlign { flow } => (flow, 0x80),
+            Self::Normal(n) => return encoder.encode(n),
+        };
+
+        encoder
+            .first_uncommitted_chunk::<1>()
+            .map(|h| h[0] = ((flow & 0x3) << 5) | extend)
     }
 }
 
@@ -152,6 +205,28 @@ impl<P> Normal<P> {
     }
 }
 
+impl<'d, U> TryFrom<Normal<decoder::Scoped<'_, 'd, U>>>
+    for Normal<payload::Payload<U::IOptions, U::DOptions>>
+where
+    U: unit::Unit,
+    U::IOptions: Encode<'d, U>,
+    U::DOptions: Encode<'d, U>,
+{
+    type Error = Error;
+
+    fn try_from(normal: Normal<decoder::Scoped<'_, 'd, U>>) -> Result<Self, Self::Error> {
+        let flow = normal.flow();
+        let src_id = normal.src_id();
+        let timestamp = normal.timestamp();
+        let res = Self::new(flow, src_id, normal.decode_payload()?);
+        if let Some(timestamp) = timestamp {
+            Ok(res.with_timestamp(timestamp))
+        } else {
+            Ok(res)
+        }
+    }
+}
+
 impl<'a, 'd, U: unit::Unit> Normal<decoder::Scoped<'a, 'd, U>> {
     /// Decode the packet's E-Trace payload
     pub fn decode_payload(mut self) -> Result<payload::Payload<U::IOptions, U::DOptions>, Error> {
@@ -162,5 +237,49 @@ impl<'a, 'd, U: unit::Unit> Normal<decoder::Scoped<'a, 'd, U>> {
             1 => Ok(payload::Payload::DataTrace),
             unknown => Err(Error::UnknownTraceType(unknown)),
         }
+    }
+}
+
+impl<'d, U> Encode<'d, U> for Normal<payload::Payload<U::IOptions, U::DOptions>>
+where
+    U: unit::Unit,
+    U::IOptions: Encode<'d, U>,
+    U::DOptions: Encode<'d, U>,
+{
+    fn encode(&self, encoder: &mut Encoder<'d, U>) -> Result<(), Error> {
+        let head = &mut encoder.first_uncommitted_chunk::<1>()?[0];
+
+        let mut original_uncommitted = encoder
+            .uncommitted()
+            .checked_sub((encoder.hart_index_width() >> 3).into())
+            .ok_or(Error::BufferTooSmall)?;
+        encoder.write_bits(self.src_id(), encoder.hart_index_width())?;
+        if let Some(timestamp) = self.timestamp() {
+            original_uncommitted = original_uncommitted
+                .checked_sub(encoder.timestamp_width().into())
+                .ok_or(Error::BufferTooSmall)?;
+            encoder.write_bits(timestamp, 8 * encoder.timestamp_width())?;
+        }
+
+        match self.payload() {
+            payload::Payload::InstructionTrace(p) => {
+                encoder.write_bits(0u8, encoder.trace_type_width())?;
+                encoder.encode(p)?;
+            }
+            payload::Payload::DataTrace => {
+                encoder.write_bits(1u8, encoder.trace_type_width())?;
+            }
+        }
+
+        let len = original_uncommitted - encoder.uncommitted();
+        let len: u8 = len
+            .try_into()
+            .ok()
+            .filter(|l| *l < 32)
+            .ok_or(Error::PayloadTooBig(len))?;
+        let flow = (self.flow() & 0x3) << 5;
+        let extend = if self.timestamp().is_some() { 0x80 } else { 0 };
+        *head = len | flow | extend;
+        Ok(())
     }
 }